refactor: refatora os mappers de usuarios para melhor definição
