+++ src/main/java/com/victorsaraiva/auth_base_jwt/dtos/user/UserResponseDTO.java
@@ -1,17 +1,18 @@
 package com.victorsaraiva.auth_base_jwt.dtos.user;
 
 import com.victorsaraiva.auth_base_jwt.models.enums.Role;
-import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.UUID;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class UserDTO {
+public class UserResponseDTO {
   private UUID id;
 
   private String username;
