+++ src/main/java/com/victorsaraiva/auth_base_jwt/mappers/impl/user/RegisterUserRequestMapperImpl.java
@@ -1,6 +1,6 @@
 package com.victorsaraiva.auth_base_jwt.mappers.impl.user;
 
-import com.victorsaraiva.auth_base_jwt.dtos.user.CreateUserDTO;
+import com.victorsaraiva.auth_base_jwt.dtos.auth.RegisterUserRequestDTO;
 import com.victorsaraiva.auth_base_jwt.mappers.Mapper;
 import com.victorsaraiva.auth_base_jwt.models.UserEntity;
 import lombok.AllArgsConstructor;
@@ -9,16 +9,16 @@ import org.springframework.stereotype.Component;
 
 @Component
 @AllArgsConstructor
-public class CreateUserMapperImpl implements Mapper<UserEntity, CreateUserDTO> {
+public class RegisterUserRequestMapperImpl implements Mapper<UserEntity, RegisterUserRequestDTO> {
   private final ModelMapper modelMapper;
 
   @Override
-  public CreateUserDTO mapTo(UserEntity userEntity) {
-    return modelMapper.map(userEntity, CreateUserDTO.class);
+  public RegisterUserRequestDTO mapTo(UserEntity userEntity) {
+    return modelMapper.map(userEntity, RegisterUserRequestDTO.class);
   }
 
   @Override
-  public UserEntity mapFrom(CreateUserDTO userDTO) {
+  public UserEntity mapFrom(RegisterUserRequestDTO userDTO) {
     return modelMapper.map(userDTO, UserEntity.class);
   }
 }
