+++ src/main/java/com/victorsaraiva/auth_base_jwt/mappers/impl/user/UserResponseMapperImpl.java
@@ -1,6 +1,6 @@
 package com.victorsaraiva.auth_base_jwt.mappers.impl.user;
 
-import com.victorsaraiva.auth_base_jwt.dtos.user.UserDTO;
+import com.victorsaraiva.auth_base_jwt.dtos.user.UserResponseDTO;
 import com.victorsaraiva.auth_base_jwt.mappers.Mapper;
 import com.victorsaraiva.auth_base_jwt.models.UserEntity;
 import lombok.AllArgsConstructor;
@@ -9,16 +9,16 @@ import org.springframework.stereotype.Component;
 
 @Component
 @AllArgsConstructor
-public class UserMapperImpl implements Mapper<UserEntity, UserDTO> {
+public class UserResponseMapperImpl implements Mapper<UserEntity, UserResponseDTO> {
   private final ModelMapper modelMapper;
 
   @Override
-  public UserDTO mapTo(UserEntity userEntity) {
-    return modelMapper.map(userEntity, UserDTO.class);
+  public UserResponseDTO mapTo(UserEntity userEntity) {
+    return modelMapper.map(userEntity, UserResponseDTO.class);
   }
 
   @Override
-  public UserEntity mapFrom(UserDTO userDTO) {
-    return modelMapper.map(userDTO, UserEntity.class);
+  public UserEntity mapFrom(UserResponseDTO userResponseDTO) {
+    return modelMapper.map(userResponseDTO, UserEntity.class);
   }
 }
